classifier/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
python-dotenv==1.0.0
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
pii-extract-plg-regex==0.5.1